requirements.txt
PyMuPDF==1.23.8
matplotlib==3.7.2
reportlab==4.0.6
google-generativeai==0.3.2
groq==0.9.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
PyPDF2==3.0.1
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.2
sentencepiece==0.1.99
protobuf==4.25.1
plotly==5.17.0
kaleido==0.2.1
seaborn==0.12.2